Changed type of Amounts from Double to float and changed details of UI

+++ Hami/DataModel/DataController.swift
@@ -28,21 +28,21 @@ class DataController: ObservableObject {
         }
     }
     
-    func addPersons(name: String, amounts: Double, comment: String, context: NSManagedObjectContext) {
+    func addPersons(name: String, amounts: Int, comment: String, context: NSManagedObjectContext) {
         let persons = Persons(context: context)
         persons.id = UUID()
         persons.date = Date()
         persons.name = name
-        persons.amounts = amounts
+        persons.amounts = Float(Double(amounts))
         persons.comment = comment
         
         save(context: context)
     }
     
-    func editPersons(persons: Persons, name: String, comment: String, amounts: Double, context: NSManagedObjectContext) {
+    func editPersons(persons: Persons, name: String, comment: String, amounts: Int, context: NSManagedObjectContext) {
         persons.date = Date()
         persons.name = name
-        persons.amounts = amounts
+        persons.amounts = Float(Double(amounts))
         persons.comment = comment
         
         save(context: context)
